fix: 댓글 및 대댓글 오류 수정
